Refactor sprites.sprites to sprites
