--- foundry/core/sprites/__init__.py
@@ -1,3 +1,0 @@
-from foundry.core.geometry import Size
-
-SPRITE_SIZE: Size = Size(8, 16)
